face_retouch_ai/requirements.txt
torch
torchvision
opencv-contrib-python==4.10.0.84
numpy<2
mediapipe>=0.10.14
protobuf==4.25.3
simple-lama-inpainting
albumentations
gdown
Pillow>=10.1
gradio
transformers
safetensors
insightface
onnxruntime
requests
scikit-image
tqdm
lpips
